2. Pre-final changes
