--- Unsplash Prototype/0. Application/SceneDelegate.swift
@@ -1,20 +1,0 @@
-
-import UIKit
-
-class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-
-    var window: UIWindow?
-    var applicationCoordinator: AppCoordinator?
-
-    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-        guard let windowScene = (scene as? UIWindowScene) else { return }
-        let window = UIWindow(windowScene: windowScene)
-        self.window = window
-        
-        let applicationCoordinator = AppCoordinator(window: window)
-        self.applicationCoordinator = applicationCoordinator
-        
-        applicationCoordinator.start()
-    }
-}
-
