--- Unsplash Prototype/1. Network Layer/UnsplashError.swift
@@ -1,12 +1,0 @@
-
-import Foundation
-
-enum UnsplashError: Error {
-    case invalidUrl
-    case noData
-    case fetchingFailure
-    
-    case decodingFailure
-    case directoryNotFound
-    case savingError
-}
