--- Unsplash Prototype/6. Coordinator/Coordinator.swift
@@ -1,17 +1,0 @@
-
-import UIKit
-
-protocol Coordinator: AnyObject {
-    var childCoordinators: [Coordinator] { get set }
-    func start()
-}
-
-extension Coordinator {
-    func addNew(coordinator: Coordinator) {
-        childCoordinators.append(coordinator)
-    }
-    
-    func removeCurrent(coordinator: Coordinator) {
-        childCoordinators = childCoordinators.filter {$0 !== coordinator }
-    }
-}
