+++ Unsplash Prototype/Application/SceneDelegate.swift
@@ -1,0 +1,18 @@
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+    var window: UIWindow?
+    var appCoordinator: AppCoordinator?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        let window = UIWindow(windowScene: windowScene)
+        self.window = window
+        
+        let applicationCoordinator = AppCoordinator(window: window)
+        self.appCoordinator = applicationCoordinator
+        
+        applicationCoordinator.start()
+    }
+}
