+++ Unsplash Prototype/Coordinator/BaseCoordinator.swift
