+++ Unsplash Prototype/Helpers/ErrorModel.swift
@@ -1,0 +1,11 @@
+
+import Foundation
+
+enum ErrorModel: Error {
+    case invalidUrl
+    case noData
+    case fetchingFailure
+    case decodingFailure
+    case directoryNotFound
+    case savingError
+}
