+++ additional1.rtf
@@ -1,0 +1,72 @@
+{\rtf1\ansi\ansicpg1251\cocoartf2818
+\cocoatextscaling0\cocoaplatform0{\fonttbl\f0\fswiss\fcharset0 Helvetica;\f1\fswiss\fcharset0 Helvetica-Bold;\f2\fnil\fcharset0 Menlo-Bold;
+\f3\fnil\fcharset0 Menlo-Regular;\f4\fnil\fcharset0 HelveticaNeue-Bold;\f5\fnil\fcharset0 HelveticaNeue;
+}
+{\colortbl;\red255\green255\blue255;\red252\green95\blue163;\red31\green31\blue36;\red255\green255\blue255;
+\red93\green216\blue255;\red115\green167\blue78;}
+{\*\expandedcolortbl;;\csgenericrgb\c98839\c37355\c63833;\csgenericrgb\c12054\c12284\c14131;\csgenericrgb\c100000\c100000\c100000\c85000;
+\csgenericrgb\c36295\c84643\c99897;\csgenericrgb\c44940\c65475\c30460;}
+\paperw11900\paperh16840\margl1440\margr1440\vieww11520\viewh8400\viewkind0
+\pard\tx566\tx1133\tx1700\tx2267\tx2834\tx3401\tx3968\tx4535\tx5102\tx5669\tx6236\tx6803\pardirnatural\partightenfactor0
+
+\f0\fs24 \cf0 \
+
+\f1\b \uc0\u1047 \u1072 \u1084 \u1077 \u1090 \u1082 \u1080 \
+\
+1. 
+\f2 \cf2 \cb3 extension
+\f3\b0 \cf4  \cf5 UnsplashLayoutCell\cf4  \{\
+\pard\tx593\pardeftab593\partightenfactor0
+\cf4     \
+    \cf6 // \uc0\u1052 \u1077 \u1090 \u1086 \u1076  \u1076 \u1083 \u1103  \u1087 \u1088 \u1086 \u1074 \u1077 \u1088 \u1082 \u1080  \u1082 \u1101 \u1096 \u1080 \u1088 \u1074 \u1086 \u1072 \u1085 \u1080 \u1103  \u1082 \u1072 \u1088 \u1090 \u1080 \u1085 \u1086 \u1082  \u1089  \u1083 \u1086 \u1075 \u1080 \u1088 \u1086 \u1074 \u1072 \u1085 \u1080 \u1077 \u1084 \cf4 \
+    \
+    \cf6 //    func configure(with image: UnsplashImage, size: ImageSize) \{\cf4 \
+    \cf6 //        let imageName = image.id\cf4 \
+    \cf6 //        let urlString = image.urls.url(for: size)\cf4 \
+    \cf6 //\cf4 \
+    \cf6 //        if let cachedImageName = ImageRepository.shared.checkIfCachedImage(for: imageName) \{\cf4 \
+    \cf6 //            if let cacheDirectory = UnsplashFileManager.cacheDirectory \{\cf4 \
+    \cf6 //                let filePath = cacheDirectory.appendingPathComponent("\\(cachedImageName).jpg")\cf4 \
+    \cf6 //                if let data = try? Data(contentsOf: filePath), let cachedImage = UIImage(data: data) \{\cf4 \
+    \cf6 //                    print("\uc0\u1048 \u1079 \u1086 \u1073 \u1088 \u1072 \u1078 \u1077 \u1085 \u1080 \u1077  \u1079 \u1072 \u1075 \u1088 \u1091 \u1078 \u1077 \u1085 \u1086  \u1080 \u1079  \u1082 \u1101 \u1096 \u1072 : \\(cachedImageName)")\cf4 \
+    \cf6 //                    DispatchQueue.main.async \{\cf4 \
+    \cf6 //                        self.imageView.image = cachedImage\cf4 \
+    \cf6 //                    \}\cf4 \
+    \cf6 //                    return\cf4 \
+    \cf6 //                \}\cf4 \
+    \cf6 //            \}\cf4 \
+    \cf6 //        \}\cf4 \
+    \cf6 //\cf4 \
+    \cf6 //        print("\uc0\u1055 \u1086 \u1087 \u1099 \u1090 \u1082 \u1072  \u1079 \u1072 \u1075 \u1088 \u1091 \u1079 \u1080 \u1090 \u1100  \u1080 \u1079 \u1086 \u1073 \u1088 \u1072 \u1078 \u1077 \u1085 \u1080 \u1077  \u1087 \u1086  URL: \\(urlString)")\cf4 \
+    \cf6 //\cf4 \
+    \cf6 //        guard let url = URL(string: urlString) else \{ return \}\cf4 \
+    \cf6 //\cf4 \
+    \cf6 //        DispatchQueue.global().async \{\cf4 \
+    \cf6 //            if let data = try? Data(contentsOf: url) \{\cf4 \
+    \cf6 //                DispatchQueue.main.async \{\cf4 \
+    \cf6 //                    print("\uc0\u1048 \u1079 \u1086 \u1073 \u1088 \u1072 \u1078 \u1077 \u1085 \u1080 \u1077  \u1079 \u1072 \u1075 \u1088 \u1091 \u1078 \u1077 \u1085 \u1086  \u1080 \u1079  \u1089 \u1077 \u1090 \u1080 : \\(urlString)")\cf4 \
+    \cf6 //                    self.imageView.image = UIImage(data: data)\cf4 \
+    \cf6 //                    ImageRepository.shared.cacheImage(from: data, for: imageName)\cf4 \
+    \cf6 //                \}\cf4 \
+    \cf6 //            \} else \{\cf4 \
+    \cf6 //                print("\uc0\u1054 \u1096 \u1080 \u1073 \u1082 \u1072  \u1079 \u1072 \u1075 \u1088 \u1091 \u1079 \u1082 \u1080  \u1080 \u1079 \u1086 \u1073 \u1088 \u1072 \u1078 \u1077 \u1085 \u1080 \u1103 : \\(urlString)")\cf4 \
+    \cf6 //            \}\cf4 \
+    \cf6 //        \}\cf4 \
+    \cf6 //    \}\cf4 \
+\}\
+\
+\
+2. 
+\f4\b\fs40 \cf0 \cb1 \uc0\u1041 \u1086 \u1083 \u1077 \u1077  \u1088 \u1072 \u1085 \u1085 \u1080 \u1081  \u1084 \u1077 \u1090 \u1086 \u1076  \u1079 \u1072 \u1075 \u1088 \u1091 \u1079 \u1082 \u1080  \u1082 \u1072 \u1088 \u1090 \u1080 \u1085 \u1086 \u1082  \u1080 \u1079  \u1055 \u1077 \u1090 -\u1087 \u1088 \u1086 \u1077 \u1082 \u1090 \u1072 \
+\pard\pardeftab560\slleading20\pardirnatural\partightenfactor0
+
+\f5\b0\fs26 \cf0 \
+\pard\pardeftab560\slleading20\partightenfactor0
+\cf0 //\
+//        DispatchQueue.global().async \{\
+//            if let data = try? Data(contentsOf: url) \{\
+//                DispatchQueue.main.async \{\
+//                    self.imageView.image = UIImage(data: data)\
+//                \}\
+//            \}\
+//        \}}
